backend/requirements.txt
Flask == 1.1.2
google-cloud-firestore == 2.1.0
google-cloud-language == 2.0.0
